Authorization functionality is done here

+++ backend/routes/posts.js
@@ -40,7 +40,10 @@ router.post(
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + "/images/" + req.file.filename
+
+    imagePath: url + "/images/" + req.file.filename,
+    creator: req.userData.userId
+
   });
   post.save().then(createdPost=>{
     res.status(201).json({
@@ -70,12 +73,17 @@ router.put(
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId
   });
-  Post.updateOne({_id: req.params.id}, post)
+  Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post)
   .then(result =>{
-    console.log(result);
-    res.status(200).json({message: "Update successfull!"});
+     console.log(result);
+    if(result.modifiedCount > 0){
+      res.status(200).json({message: "Update successfull!"});
+    } else {
+      res.status(401).json({message: "Not authorized!"});
+    }
   });
 })
 
@@ -114,10 +122,13 @@ router.get("/:id", (req, res, next)=>{
 })
 
 router.delete("/:id",checkAuth, (req, res, next)=>{
-  Post.deleteOne({_id: req.params.id})
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId})
   .then(result =>{
-    console.log(result);
-    res.status(200).json({message: "Post Del!"});
+    if(result.deletedCount > 0){
+      res.status(200).json({message: "Deletion successfull!"});
+    } else {
+      res.status(401).json({message: "Not authorized!"});
+    }
   })
 });
